requirements.txt
psycopg2-binary>=2.9.3
pyyaml>=6.0
selenium>=4.1.0
beautifulsoup4>=4.10.0
webdriver-manager>=3.5.2
pandas>=1.3.5
tabulate>=0.8.0
python-dotenv>=0.19.0 